rename module 'stats' to 'agg'
