+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import setup
 
 setup(
     name="inrix",
-    packages=["inrix", "inrix.api", "inrix.stats"],
-    version="0.beta1",
+    packages=["inrix", "inrix.api", "inrix.agg"],
+    version="0.beta2",
     description="Scripts for analyzing INRIX IQ Roadway Analytics data",
 )
